feat(dap): evaluate request via shared expression compiler
Adds DAP evaluate so the editor's watch panel, hover tooltip, and
debug console all see the same expression grammar that powers
\`:bp X if E\` and \`:bpr X if E\` in the TUI.

The compiler moves from internal/tui/cond.go to a new internal/expr
package so internal/dap can use it without depending on the TUI.
Public API:

  expr.EvalFn = func(*cpu.CPU, cpu.Bus) uint32
  expr.Compile(src, syms) (EvalFn, error)

tui.compileCondition is now a thin wrapper that maps "non-zero" to
true. Behavior unchanged; all existing TUI conditional-bp tests
continue to pass against the relocated grammar.

DAP evaluate compiles the request expression, evaluates it against
the live CPU + bus, and returns the result formatted by natural
width: \$XX for byte-sized, \$XXXX for word-sized, \$XXXXXXXX otherwise.

Closes #52.
